refactor: extract target selector from ruby toolbar
Move target navigation (prev/next sprite), command input search,
and dropdown logic into a separate TargetSelector component.
ruby-toolbar.jsx: 542 → 275 lines.

Part of #275 (Phase 4)
