+++ packages/scratch-gui/src/components/ruby-toolbar/ruby-toolbar.jsx
@@ -5,14 +5,13 @@ import VM from '@smalruby/scratch-vm';
 
 import styles from './ruby-toolbar.css';
 import messages from './messages.js';
+import TargetSelector from './target-selector.jsx';
 
 import iconPlay from './icon--play.svg';
 import iconStop from './icon--stop.svg';
 import iconSearch from './icon--search.svg';
 import iconUndo from './icon--undo.svg';
 import iconRedo from './icon--redo.svg';
-import iconBack from './icon--back.svg';
-import iconForward from './icon--forward.svg';
 import iconDownload from './icon--download.svg';
 import iconAI from './icon--ai.svg';
 import iconFurigana from './icon--furigana.svg';
@@ -20,9 +19,6 @@ import iconAutoCorrect from './icon--auto-correct.svg';
 
 const RubyToolbar = props => {
     const intl = useIntl();
-    const [commandValue, setCommandValue] = useState('');
-    const [showDropdown, setShowDropdown] = useState(false);
-    const [filteredTargets, setFilteredTargets] = useState([]);
     const [showMoreMenu, setShowMoreMenu] = useState(false);
     const moreMenuRef = useRef(null);
 
@@ -41,212 +37,46 @@ const RubyToolbar = props => {
         };
     }, [showMoreMenu]);
 
-    const getSortedSprites = useCallback(() => {
-        if (!props.vm || !props.vm.runtime) {
-            return [];
-        }
-        const targets = props.vm.runtime.targets;
-        const sprites = targets
-            .filter(t => !t.isStage)
-            .sort((a, b) => a.getLayerOrder() - b.getLayerOrder());
-        return sprites;
-    }, [props.vm]);
-
-    const getAllTargets = useCallback(() => {
-        if (!props.vm || !props.vm.runtime) {
-            return [];
-        }
-        const stage = props.vm.runtime.targets.find(t => t.isStage);
-        const sprites = getSortedSprites();
-        return stage ? [stage, ...sprites] : sprites;
-    }, [props.vm, getSortedSprites]);
-
-    const getCurrentSpriteIndex = useCallback(() => {
-        const sprites = getSortedSprites();
-        const currentId = props.editingTarget?.id;
-        return sprites.findIndex(s => s.id === currentId);
-    }, [getSortedSprites, props.editingTarget]);
-
-    const getTargetName = useCallback(target => {
-        if (!target) {
-            return '';
-        }
-        if (target.isStage) {
-            return intl.formatMessage(messages.stage);
-        }
-        if (typeof target.getName === 'function') {
-            return target.getName();
-        }
-        if (target.sprite && target.sprite.name) {
-            return target.sprite.name;
-        }
-        return target.name || '';
-    }, [intl]);
-
     const handleSearch = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
+        if (props.onDismissBubble) props.onDismissBubble();
         if (props.editorRef) {
-            // Trigger Monaco Editor's search action
             props.editorRef.trigger('keyboard', 'actions.find', null);
         }
     }, [props]);
 
     const handleUndo = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
+        if (props.onDismissBubble) props.onDismissBubble();
         if (props.editorRef) {
             props.editorRef.trigger('keyboard', 'undo', null);
         }
     }, [props]);
 
     const handleRedo = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
+        if (props.onDismissBubble) props.onDismissBubble();
         if (props.editorRef) {
             props.editorRef.trigger('keyboard', 'redo', null);
         }
     }, [props]);
 
-    const handlePrevSprite = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
-
-        const sprites = getSortedSprites();
-        const currentIndex = getCurrentSpriteIndex();
-
-        if (props.editingTarget?.isStage) {
-            return;
-        }
-
-        if (currentIndex === 0) {
-            const stage = props.vm.runtime.targets.find(t => t.isStage);
-            if (stage) {
-                props.onSelectTarget(stage.id);
-            }
-        } else if (currentIndex > 0) {
-            props.onSelectTarget(sprites[currentIndex - 1].id);
-        }
-    }, [getSortedSprites, getCurrentSpriteIndex, props]);
-
-    const handleNextSprite = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
-
-        const sprites = getSortedSprites();
-        const currentIndex = getCurrentSpriteIndex();
-
-        if (props.editingTarget?.isStage) {
-            if (sprites.length > 0) {
-                props.onSelectTarget(sprites[0].id);
-            }
-        } else if (currentIndex >= 0 && currentIndex < sprites.length - 1) {
-            props.onSelectTarget(sprites[currentIndex + 1].id);
-        }
-    }, [getSortedSprites, getCurrentSpriteIndex, props]);
-
-    const handleCommandChange = useCallback(e => {
-        const value = e.target.value;
-        setCommandValue(value);
-
-        if (value.startsWith('>')) {
-            // Command mode - TODO: implement Monaco command palette
-            setShowDropdown(false);
-        } else {
-            // Sprite search mode
-            const allTargets = getAllTargets();
-            const filtered = allTargets.filter(target => {
-                const name = getTargetName(target).toLowerCase();
-                return name.includes(value.toLowerCase());
-            });
-            setShowDropdown(value.length > 0 && filtered.length > 0);
-            setFilteredTargets(filtered);
-        }
-    }, [getAllTargets, getTargetName]);
-
-    const handleCommandFocus = useCallback(() => {
-        if (commandValue && !commandValue.startsWith('>')) {
-            const allTargets = getAllTargets();
-            const filtered = allTargets.filter(target => {
-                const name = getTargetName(target).toLowerCase();
-                return name.includes(commandValue.toLowerCase());
-            });
-            setShowDropdown(filtered.length > 0);
-            setFilteredTargets(filtered);
-        }
-    }, [commandValue, getAllTargets, getTargetName]);
-
-    const handleCommandBlur = useCallback(() => {
-        // Delay to allow click on dropdown item
-        setTimeout(() => {
-            setShowDropdown(false);
-        }, 200);
-    }, []);
-
-    const handleSelectTarget = useCallback(targetId => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
-        props.onSelectTarget(targetId);
-        setCommandValue('');
-        setShowDropdown(false);
-    }, [props]);
-
-    const handleCommandKeyDown = useCallback(e => {
-        if (e.key === 'Escape') {
-            setCommandValue('');
-            setShowDropdown(false);
-            e.target.blur();
-        } else if (e.key === 'Enter' && !e.isComposing && filteredTargets.length > 0) {
-            handleSelectTarget(filteredTargets[0].id);
-        }
-    }, [filteredTargets, handleSelectTarget]);
-
-    const handleSelectTargetFromDropdown = useCallback(e => {
-        const targetId = e.currentTarget.dataset.targetId;
-        handleSelectTarget(targetId);
-    }, [handleSelectTarget]);
-
     const handleDownload = useCallback(() => {
         setShowMoreMenu(false);
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
-        if (props.onDownload) {
-            props.onDownload();
-        }
+        if (props.onDismissBubble) props.onDismissBubble();
+        if (props.onDownload) props.onDownload();
     }, [props]);
 
     const handleOpenAI = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
-        if (props.onOpenGeminiModal) {
-            props.onOpenGeminiModal();
-        }
+        if (props.onDismissBubble) props.onDismissBubble();
+        if (props.onOpenGeminiModal) props.onOpenGeminiModal();
     }, [props]);
 
     const handleToggleFurigana = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
-        if (props.onToggleFurigana) {
-            props.onToggleFurigana();
-        }
+        if (props.onDismissBubble) props.onDismissBubble();
+        if (props.onToggleFurigana) props.onToggleFurigana();
     }, [props]);
 
     const handleToggleAutoCorrect = useCallback(() => {
-        if (props.onDismissBubble) {
-            props.onDismissBubble();
-        }
-        if (props.onToggleAutoCorrect) {
-            props.onToggleAutoCorrect();
-        }
+        if (props.onDismissBubble) props.onDismissBubble();
+        if (props.onToggleAutoCorrect) props.onToggleAutoCorrect();
     }, [props]);
 
     const handleToggleMoreMenu = useCallback(() => {
@@ -255,61 +85,14 @@ const RubyToolbar = props => {
 
     const handleOpenAutoCorrectSettings = useCallback(() => {
         setShowMoreMenu(false);
-        if (props.onOpenAutoCorrectSettings) {
-            props.onOpenAutoCorrectSettings();
-        }
+        if (props.onOpenAutoCorrectSettings) props.onOpenAutoCorrectSettings();
     }, [props]);
 
     const handleExecuteLine = useCallback(() => {
-        if (!props.editorRef) {
-            return;
-        }
-
-        // Get current cursor position
+        if (!props.editorRef) return;
         const position = props.editorRef.getPosition();
-        const lineNumber = position.lineNumber;
-
-        // Trigger conversion and execution
-        if (props.onExecuteLine) {
-            props.onExecuteLine(lineNumber);
-        }
+        if (props.onExecuteLine) props.onExecuteLine(position.lineNumber);
     }, [props]);
-
-    const canGoPrev = useCallback(() => {
-        if (props.editingTarget?.isStage) {
-            return false;
-        }
-        const currentIndex = getCurrentSpriteIndex();
-        return currentIndex >= 0;
-    }, [props.editingTarget, getCurrentSpriteIndex]);
-
-    const canGoNext = useCallback(() => {
-        const sprites = getSortedSprites();
-        if (sprites.length === 0) {
-            return false;
-        }
-        if (props.editingTarget?.isStage) {
-            return true;
-        }
-        const currentIndex = getCurrentSpriteIndex();
-        return currentIndex >= 0 && currentIndex < sprites.length - 1;
-    }, [getSortedSprites, props.editingTarget, getCurrentSpriteIndex]);
-
-    const highlightMatch = useCallback((text, query) => {
-        if (!query) return text;
-        const index = text.toLowerCase().indexOf(query.toLowerCase());
-        if (index === -1) return text;
-
-        return (
-            <>
-                {text.substring(0, index)}
-                <span className={styles.highlight}>
-                    {text.substring(index, index + query.length)}
-                </span>
-                {text.substring(index + query.length)}
-            </>
-        );
-    }, []);
 
     return (
         <div className={styles.toolbar}>
@@ -319,8 +102,12 @@ const RubyToolbar = props => {
                     className={styles.iconButton}
                     onClick={handleExecuteLine}
                     disabled={!props.editorRef}
-                    aria-label={intl.formatMessage(props.isRunning ? messages.stopExecution : messages.executeLine)}
-                    title={intl.formatMessage(props.isRunning ? messages.stopExecution : messages.executeLine)}
+                    aria-label={intl.formatMessage(
+                        props.isRunning ? messages.stopExecution : messages.executeLine
+                    )}
+                    title={intl.formatMessage(
+                        props.isRunning ? messages.stopExecution : messages.executeLine
+                    )}
                 >
                     <img
                         src={props.isRunning ? iconStop : iconPlay}
@@ -374,11 +161,17 @@ const RubyToolbar = props => {
             {/* Furigana Toggle & Auto Correct Toggle */}
             <div className={`${styles.toolbarPart} ${styles.modDashedBorder}`}>
                 <button
-                    className={`${styles.furiganaButton} ${props.furiganaEnabled ? styles.furiganaButtonActive : ''}`}
+                    className={`${styles.furiganaButton} ${
+                        props.furiganaEnabled ? styles.furiganaButtonActive : ''
+                    }`}
                     onClick={handleToggleFurigana}
-                    aria-label={intl.formatMessage(props.furiganaEnabled ? messages.furiganaOn : messages.furiganaOff)}
+                    aria-label={intl.formatMessage(
+                        props.furiganaEnabled ? messages.furiganaOn : messages.furiganaOff
+                    )}
                     aria-pressed={props.furiganaEnabled}
-                    title={intl.formatMessage(props.furiganaEnabled ? messages.furiganaOn : messages.furiganaOff)}
+                    title={intl.formatMessage(
+                        props.furiganaEnabled ? messages.furiganaOn : messages.furiganaOff
+                    )}
                 >
                     <img
                         src={iconFurigana}
@@ -407,56 +200,12 @@ const RubyToolbar = props => {
 
             {/* Navigation & Command Part */}
             <div className={`${styles.toolbarPart} ${styles.modDashedBorder} ${styles.modCenter}`}>
-                <button
-                    className={styles.iconButton}
-                    onClick={handlePrevSprite}
-                    disabled={!canGoPrev()}
-                    aria-label={intl.formatMessage(messages.prevSprite)}
-                    title={intl.formatMessage(messages.prevSprite)}
-                >
-                    <img
-                        src={iconBack}
-                        alt=""
-                    />
-                </button>
-                <button
-                    className={styles.iconButton}
-                    onClick={handleNextSprite}
-                    disabled={!canGoNext()}
-                    aria-label={intl.formatMessage(messages.nextSprite)}
-                    title={intl.formatMessage(messages.nextSprite)}
-                >
-                    <img
-                        src={iconForward}
-                        alt=""
-                    />
-                </button>
-                <div className={styles.commandWrapper}>
-                    <input
-                        className={styles.commandInput}
-                        type="text"
-                        value={commandValue}
-                        onChange={handleCommandChange}
-                        onFocus={handleCommandFocus}
-                        onBlur={handleCommandBlur}
-                        onKeyDown={handleCommandKeyDown}
-                        placeholder={intl.formatMessage(messages.commandPlaceholder)}
-                    />
-                    {showDropdown && (
-                        <div className={styles.dropdown}>
-                            {filteredTargets.map(target => (
-                                <div
-                                    key={target.id}
-                                    className={styles.dropdownItem}
-                                    data-target-id={target.id}
-                                    onMouseDown={handleSelectTargetFromDropdown}
-                                >
-                                    {highlightMatch(getTargetName(target), commandValue)}
-                                </div>
-                            ))}
-                        </div>
-                    )}
-                </div>
+                <TargetSelector
+                    editingTarget={props.editingTarget}
+                    vm={props.vm}
+                    onSelectTarget={props.onSelectTarget}
+                    onDismissBubble={props.onDismissBubble}
+                />
                 <button
                     className={styles.iconButton}
                     onClick={handleOpenAI}
@@ -495,9 +244,7 @@ const RubyToolbar = props => {
                                     src={iconDownload}
                                     alt=""
                                 />
-                                {intl.formatMessage(
-                                    messages.saveRubyScript
-                                )}
+                                {intl.formatMessage(messages.saveRubyScript)}
                             </div>
                             <div
                                 className={styles.moreMenuItem}
@@ -508,9 +255,7 @@ const RubyToolbar = props => {
                                     src={iconAutoCorrect}
                                     alt=""
                                 />
-                                {intl.formatMessage(
-                                    messages.autoCorrectSettings
-                                )}
+                                {intl.formatMessage(messages.autoCorrectSettings)}
                             </div>
                         </div>
                     )}
